sirono (#4)
Changed CollectionUtilTest so that it no longer uses Account. This was causing issues when we were using it directly because of a unique custom field in our own metadata.

Made it possible to subclass PicklistEnum and TypeSafeEnum, primarily so that new intermediate base classes could be created that allow these enum types and their constants to be exported from managed packages.